requirements.txt
astroid==1.5.3
backports.functools-lru-cache==1.4
configparser==3.5.0
crlibm==1.0.3
cycler==0.10.0
enum34==1.1.6
ete3==3.0.0b35
functools32==3.2.3.post2
isort==4.2.15
lazy-object-proxy==1.3.1
matplotlib==2.0.2
mccabe==0.6.1
numpy==1.13.1
pyinterval==1.2.0
pylint==1.7.2
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
ruamel.ordereddict==0.4.13
ruamel.yaml==0.15.31
scipy==0.19.1
singledispatch==3.4.0.3
six==1.10.0
subprocess32==3.2.7
vext==0.6.9
vext.vtk==0.6.8
wrapt==1.10.11
